% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTestData-method.R
\name{getTestData}
\alias{getTestData}
\alias{getTestData,LncRNAData-method}
\title{Get Test Data}
\usage{
getTestData(object)

\S4method{getTestData}{LncRNAData}(object)
}
\arguments{
\item{object}{An object of class \code{LncRNAData}.}
}
\value{
A data.frame containing the test data.
}
\description{
This method retrieves the test data stored in the package.
}
